Unwind _TestTypes to top-level public test classes in AOP test suite
(cherry picked from commit cdaa247)
