+++ spring-context/src/test/java/org/springframework/aop/aspectj/AdviceBindingTestAspect.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2012 the original author or authors.
+ * Copyright 2002-2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,23 +19,6 @@ package org.springframework.aop.aspectj;
 import org.aspectj.lang.JoinPoint;
 
 /**
- * Definitions of testing types for use in within this package.
- * Wherever possible, test types should be defined local to the java
- * file that makes use of them.  In some cases however, a test type may
- * need to be shared across tests.  Such types reside here, with the
- * intention of reducing the surface area of java files within this
- * package.  This allows developers to think about tests first, and deal
- * with these second class testing artifacts on an as-needed basis.
- *
- * Types here should be defined as package-private top level classes in
- * order to avoid needing to fully qualify, e.g.: _TestTypes$Foo.
- *
- * @author Chris Beams
- */
-final class _TestTypes { }
-
-
-/**
  * Aspect used as part of before advice binding tests and
  * serves as base class for a number of more specialized test aspects.
  *
@@ -44,13 +27,16 @@ final class _TestTypes { }
  */
 class AdviceBindingTestAspect {
 
-	protected AdviceBindingCollaborator collaborator = null;
+	protected AdviceBindingCollaborator collaborator;
+
 
 	public void setCollaborator(AdviceBindingCollaborator aCollaborator) {
 		this.collaborator = aCollaborator;
 	}
 
+
 	// "advice" methods
+
 	public void oneIntArg(int age) {
 		this.collaborator.oneIntArg(age);
 	}
@@ -79,67 +65,14 @@ class AdviceBindingTestAspect {
 	public interface AdviceBindingCollaborator {
 
 		void oneIntArg(int x);
+
 		void oneObjectArg(Object o);
+
 		void oneIntAndOneObject(int x, Object o);
+
 		void needsJoinPoint(String s);
+
 		void needsJoinPointStaticPart(String s);
 	}
 
-}
-
-
-/**
- * @author Ramnivas Laddad
- */
-interface ICounter {
-
-	void increment();
-
-	void decrement();
-
-	int getCount();
-
-	void setCount(int counter);
-
-	void reset();
-
-}
-
-
-/**
- * A simple counter for use in simple tests (for example, how many times an advice was executed)
- *
- * @author Ramnivas Laddad
- */
-final class Counter implements ICounter {
-
-	private int count;
-
-	public Counter() {
-	}
-
-	@Override
-	public void increment() {
-		count++;
-	}
-
-	@Override
-	public void decrement() {
-		count--;
-	}
-
-	@Override
-	public int getCount() {
-		return count;
-	}
-
-	@Override
-	public void setCount(int counter) {
-		this.count = counter;
-	}
-
-	@Override
-	public void reset() {
-		this.count = 0;
-	}
 }
