+++ spring-context/src/test/java/org/springframework/aop/framework/Echo.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2002-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.aop.framework;
+
+public class Echo implements IEcho {
+
+	private int a;
+
+	@Override
+	public int echoException(int i, Throwable t) throws Throwable {
+		if (t != null) {
+			throw t;
+		}
+		return i;
+	}
+
+	@Override
+	public void setA(int a) {
+		this.a = a;
+	}
+
+	@Override
+	public int getA() {
+		return a;
+	}
+
+}
